Refactor architetturale enterprise: performance, Decimal, moduli, Swift 6, test
Performance:
- Account.currentBalance/futureBalance da O(n) per render a cache O(1)
  (AccountBalanceCache, invalidata da ModelContext.safeSave())
- processRecurring spostato su RecurringTransactionActor (@ModelActor background)
  invece di girare sincrono sul main actor al lancio
- BudgetHistorySheet.outflowsByMonth precalcolato una volta in init invece che
  ad ogni render

Double -> Decimal: importi monetari (Account/Transaction/Budget/Goal) migrati
a Decimal con VersionedSchema (SchemaV1/SchemaV2) + MigrationPlan custom per
convertire senza perdita gli store esistenti; formatter, parsing e DTO Supabase
aggiornati di conseguenza.

Modularizzazione: Models.swift e CategoryClassifier.swift (God File) spezzati
in Models/, Domain/, Formatting/, Persistence/; il dizionario keyword del
classificatore spostato in Resources/CategoryKeywords.json dietro il
protocollo CategoryClassifying (KeywordCategoryClassifier).

Concurrency & sicurezza: SWIFT_VERSION 6.0 con strict concurrency attiva;
NotificationManager estratto dietro il protocollo NotificationScheduling
(SystemNotificationManager); nuovo gate Face ID/Touch ID/passcode
(BiometricAuthenticating + AppLockGate) con toggle in Impostazioni.

Test: nuovo target MoneyTrackerTests (Swift Testing) con copertura di
classificatore, calcolo saldi Decimal e migrazione schema V1->V2 end-to-end.

+++ MoneyTracker/DashboardView.swift
@@ -42,9 +42,9 @@ struct DashboardView: View {
 
     // MARK: - Totals
 
-    private var totalCurrent: Double { includedAccounts.reduce(0) { $0 + $1.currentBalance } }
-    private var totalFuture:  Double { includedAccounts.reduce(0) { $0 + $1.futureBalance } }
-    private var totalDelta:   Double { totalFuture - totalCurrent }
+    private var totalCurrent: Decimal { includedAccounts.reduce(0) { $0 + $1.currentBalance } }
+    private var totalFuture:  Decimal { includedAccounts.reduce(0) { $0 + $1.futureBalance } }
+    private var totalDelta:   Decimal { totalFuture - totalCurrent }
 
     private var recentNormal: [Transaction] {
         Array(transactions.filter { !$0.isFixed && !$0.isTransfer && $0.recurringFrequency.isEmpty }.prefix(4))
